Card commands are collection commands now.
